added tutorials package to the project.
enable user to view the list of albums.
enable user to vew details of an album.
enable user to add a new album.
enable user to update an album.
enable user to add articles to an album.
enable user to view details of an article.
enable user to update an article.

enable user to give comments to an album/article.

enable user to view their own provile.